Move logic into google_oauth
